CI: Use makepkg, which generates a package as build result
The resulting package could be uploaded to a repository, which then can
be fed back into the process that builds the base Docker images for CI.

For example: the result from building the libwpe package could be
included as part of the image used for CI of e.g. WPEBackend-fdo.